line-bot-news/requirements.txt
Flask
line-bot-sdk
openai
requests
python-dotenv
h11
